Added param to PlayerCommand and added Take command
I've added a result String to the PlayerCommand data, so that we can update the player and have a return message in the same line. I've also added the Take command so that the player can take items from his or her current location and add it to their inventory.